# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  description: REST API для управления банковскими картами
  version: 1.0.0
  contact:
    name: Bank Support
    email: support@bank.com

servers:
  - url: http://localhost:8080
    description: Development server

security:
  - bearerAuth: []

paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Email already registered or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login and get JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/cards:
    get:
      tags:
        - Cards
      summary: Get current user's cards
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CardStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardPage'
    post:
      tags:
        - Cards
      summary: Create a new card (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Card created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/cards/all:
    get:
      tags:
        - Cards
      summary: Get all cards (Admin only)
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CardStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of all cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardPage'
        '403':
          description: Access denied

  /api/cards/{id}:
    get:
      tags:
        - Cards
      summary: Get card by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Card details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Card not found
    delete:
      tags:
        - Cards
      summary: Delete card (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Card deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Access denied

  /api/cards/{id}/block:
    put:
      tags:
        - Cards
      summary: Block card (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Card blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Card already blocked

  /api/cards/{id}/activate:
    put:
      tags:
        - Cards
      summary: Activate card (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Card activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Card already active or expired

  /api/cards/{id}/request-block:
    post:
      tags:
        - Cards
      summary: Request to block own card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Card block requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '403':
          description: Not your card

  /api/cards/{id}/balance:
    get:
      tags:
        - Cards
      summary: Get card balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Card balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '403':
          description: Access denied

  /api/cards/transfer:
    post:
      tags:
        - Cards
      summary: Transfer money between own cards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Transfer completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Transfer error

  /api/users:
    get:
      tags:
        - Users
      summary: Get all users (Admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '403':
          description: Access denied

  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
    delete:
      tags:
        - Users
      summary: Delete user (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Cannot delete own account

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        type:
          type: string
        email:
          type: string
        role:
          type: string

    CreateCardRequest:
      type: object
      required:
        - ownerId
        - expiryDate
      properties:
        ownerId:
          type: integer
          format: int64
        expiryDate:
          type: string
          format: date
        initialBalance:
          type: number

    TransferRequest:
      type: object
      required:
        - fromCardId
        - toCardId
        - amount
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number

    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedCardNumber:
          type: string
        ownerId:
          type: integer
          format: int64
        ownerEmail:
          type: string
        expiryDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/CardStatus'
        balance:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    BalanceResponse:
      type: object
      properties:
        cardId:
          type: integer
          format: int64
        maskedCardNumber:
          type: string
        balance:
          type: number

    TransferResponse:
      type: object
      properties:
        message:
          type: string
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
        fromCardNewBalance:
          type: number
        toCardNewBalance:
          type: number

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    MessageResponse:
      type: object
      properties:
        message:
          type: string

    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    CardStatus:
      type: string
      enum: [ACTIVE, BLOCKED, EXPIRED]

    CardPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    UserPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer