# src/main/java/com/example/bankcards/dto/response/BalanceResponse.java
package com.example.bankcards.dto.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class BalanceResponse {

    private Long cardId;
    private String maskedCardNumber;
    private BigDecimal balance;
}